LayoutTests/http/tests/notifications/service-worker-notification-permission-expected.txt

This test checks that Notification::permission works for service workers

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS message.data is "default"
PASS receivedPostMessageResponse became true
PASS requestPermissionResult is "granted"
PASS message.data is "granted"
PASS receivedPostMessageResponse became true
PASS requestPermissionResult is "denied"
PASS message.data is "denied"
PASS receivedPostMessageResponse became true
PASS successfullyParsed is true

TEST COMPLETE
